Refactor trpc router code to teensy.ts
