--- src/server/api/routers/index.ts
@@ -1,3 +1,0 @@
-import { createTRPCRouter } from "@/server/api/trpc";
-
-export const postRouter = createTRPCRouter({});
